go: change the Go version support policy, drop one version
Our policy tried to be a bit more friendly to people, but unfortunately it's
somewhat unsynchronized with the rest of the world using Go. This can be
a problem for dependency management (including security fixes), newer
libraries require newer Go and we're still trying to allow for an older one.
Also, an additional version means more test runs and dropping one can save
some CI time.

Fixes #30.

Signed-off-by: Roman Khimov <roman@nspcc.ru>